use safeMove in project output to work across filesystem boundaries
